needifyv1/dprojx/requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
gunicorn==20.0.4
install==1.3.4
mysql-client==0.0.1
nginx==0.0.1
Pillow==8.1.0
PyMySQL==1.0.1
pytz==2020.5
sqlparse==0.4.1
